perf: lazy-load Skland credential form
