+++ scripts/check-bundle-budget.mjs
@@ -7,13 +7,13 @@ import { gzipSync } from "node:zlib";
 // The calculator keeps its always-visible board in the initial graph. Secondary workbench
 // views have independent route chunks and may carry their own datasets without joining `/`.
 const MAX_SKLAND_DISABLED_ROUTE_INITIAL_JS_BYTES = 1_130_000;
-const MAX_SKLAND_ENABLED_ROUTE_INITIAL_JS_BYTES = 1_150_000;
-const MAX_SECONDARY_ROUTE_INITIAL_JS_BYTES = 1_525_000;
-const MAX_SKLAND_ROUTE_INITIAL_JS_BYTES = 1_590_000;
+const MAX_SKLAND_ENABLED_ROUTE_INITIAL_JS_BYTES = 1_160_000;
+const MAX_SECONDARY_ROUTE_INITIAL_JS_BYTES = 1_540_000;
+const MAX_SKLAND_ROUTE_INITIAL_JS_BYTES = 1_600_000;
 const MAX_SKLAND_DISABLED_DOCUMENT_INITIAL_JS_BYTES = 1_240_000;
-const MAX_SKLAND_ENABLED_DOCUMENT_INITIAL_JS_BYTES = 1_270_000;
+const MAX_SKLAND_ENABLED_DOCUMENT_INITIAL_JS_BYTES = 1_275_000;
 const MAX_SKLAND_DISABLED_DOCUMENT_INITIAL_GZIP_JS_BYTES = 395_000;
-const MAX_SKLAND_ENABLED_DOCUMENT_INITIAL_GZIP_JS_BYTES = 405_000;
+const MAX_SKLAND_ENABLED_DOCUMENT_INITIAL_GZIP_JS_BYTES = 407_000;
 const MAX_DOCUMENT_INITIAL_JS_FILES = 18;
 const WORKBENCH_ROUTES = ["/", "/training", "/skills", "/skland", "/account"];
 const statsUrl = new URL("../.next/diagnostics/route-bundle-stats.json", import.meta.url);
