+++ src/skland-components.tsx
@@ -1,37 +1,44 @@
 "use client";
 
-import Link from "next/link";
-import { useCallback, useEffect, useId, useRef, useState, type FormEvent } from "react";
+import dynamic from "next/dynamic";
+import { useCallback, useEffect, useId, useRef, useState, type KeyboardEvent } from "react";
 import {
-  Check,
-  Clipboard,
   ExternalLink,
   KeyRound,
   LoaderCircle,
   RefreshCw,
   ScanLine,
   ShieldCheck,
-  TriangleAlert,
 } from "lucide-react";
 import { QRCodeSVG } from "qrcode.react";
 
-import { importSklandCredential, pollSklandQr, startSklandQr, toDisplayError } from "@/api";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
+import { pollSklandQr, startSklandQr, toDisplayError } from "@/api";
+import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import { Card, CardDescription, CardTitle } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { cn } from "@/lib/utils";
 import { buildSklandAppOpenUrl } from "@/skland-auth-url";
+import {
+  currentSklandPolicyConsent,
+  SklandPolicyConsent,
+} from "@/skland-policy-consent";
 import type { SklandSessionData } from "@/types";
-import { PRIVACY_VERSION, TERMS_VERSION, type SklandPolicyConsentRequest } from "@/legal-policy";
 
 const SKLAND_QR_POLL_INTERVAL_MS = 6_000;
-export const SKLAND_CREDENTIAL_COPY_COMMAND = 'copy([localStorage.getItem("SK_OAUTH_CRED_KEY"), localStorage.getItem("SK_TOKEN_CACHE_KEY")].join(","))';
+const SklandCredentialPanel = dynamic(
+  () => import("@/skland-credential-panel").then((module) => module.SklandCredentialPanel),
+  {
+    ssr: false,
+    loading: () => (
+      <div className="flex min-h-32 items-center justify-center gap-2 text-sm text-muted-foreground" role="status">
+        <LoaderCircle className="animate-spin motion-reduce:animate-none" />正在加载凭证导入…
+      </div>
+    ),
+  },
+);
 
 type ScanState = "idle" | "loading" | "waiting" | "scanned" | "expired";
 type AuthMethod = "qr" | "credential";
-type ImportState = "idle" | "submitting";
 
 interface SklandLoginPanelProps {
   configured: boolean;
@@ -39,62 +46,6 @@ interface SklandLoginPanelProps {
   onAuthenticated: (session: SklandSessionData) => void;
   className?: string;
   dialogPresentation?: boolean;
-}
-
-function policyConsent(): SklandPolicyConsentRequest {
-  return {
-    termsAccepted: true,
-    privacyAccepted: true,
-    termsVersion: TERMS_VERSION,
-    privacyVersion: PRIVACY_VERSION,
-  };
-}
-
-function PolicyConsent({
-  termsAccepted,
-  privacyAccepted,
-  onTermsChange,
-  onPrivacyChange,
-}: {
-  termsAccepted: boolean;
-  privacyAccepted: boolean;
-  onTermsChange: (accepted: boolean) => void;
-  onPrivacyChange: (accepted: boolean) => void;
-}) {
-  const termsId = useId();
-  const privacyId = useId();
-  return (
-    <div className="grid w-full gap-3 text-start text-xs leading-5 text-muted-foreground" data-skland-policy-consent>
-      <div className="flex items-start gap-2">
-        <input
-          id={termsId}
-          type="checkbox"
-          checked={termsAccepted}
-          onChange={(event) => onTermsChange(event.target.checked)}
-          className="mt-1 size-4 shrink-0 accent-primary"
-        />
-        <label htmlFor={termsId}>
-          我已阅读并同意
-          <Link className="mx-1 font-medium text-foreground underline underline-offset-4" href="/terms" target="_blank">本站服务条款</Link>
-          。
-        </label>
-      </div>
-      <div className="flex items-start gap-2">
-        <input
-          id={privacyId}
-          type="checkbox"
-          checked={privacyAccepted}
-          onChange={(event) => onPrivacyChange(event.target.checked)}
-          className="mt-1 size-4 shrink-0 accent-primary"
-        />
-        <label htmlFor={privacyId}>
-          我已阅读
-          <Link className="mx-1 font-medium text-foreground underline underline-offset-4" href="/privacy" target="_blank">本站隐私政策</Link>
-          ，并同意本站为登录和数据同步处理我的森空岛凭证、角色、干员与基建数据。
-        </label>
-      </div>
-    </div>
-  );
 }
 
 export function SklandLoginPanel({
@@ -114,11 +65,7 @@ export function SklandLoginPanel({
   const [preparingSlow, setPreparingSlow] = useState(false);
   const [termsAccepted, setTermsAccepted] = useState(false);
   const [privacyAccepted, setPrivacyAccepted] = useState(false);
-  const [credential, setCredential] = useState("");
-  const [importState, setImportState] = useState<ImportState>("idle");
-  const [importError, setImportError] = useState<string | null>(null);
-  const [copyState, setCopyState] = useState<"idle" | "copied" | "error">("idle");
-  const credentialInputId = useId();
+  const authTabsId = useId();
   const createQrPromiseRef = useRef<Promise<void> | null>(null);
   const mountedRef = useRef(true);
   const consentReady = termsAccepted && privacyAccepted;
@@ -139,7 +86,7 @@ export function SklandLoginPanel({
       setScanUrl(null);
       setScanExpiresAt(null);
       try {
-        const result = await startSklandQr(policyConsent());
+        const result = await startSklandQr(currentSklandPolicyConsent());
         if (!mountedRef.current) return;
         setScanId(result.scanId);
         setScanUrl(result.scanUrl);
@@ -251,35 +198,16 @@ export function SklandLoginPanel({
     void createQr();
   }, [authMethod, configured, consentReady, createQr, dialogPresentation, scanError, scanState]);
 
-  async function copyCredentialCommand() {
-    try {
-      await navigator.clipboard.writeText(SKLAND_CREDENTIAL_COPY_COMMAND);
-      setCopyState("copied");
-      window.setTimeout(() => {
-        if (mountedRef.current) setCopyState("idle");
-      }, 2_000);
-    } catch {
-      setCopyState("error");
-    }
-  }
-
-  async function submitCredential(event: FormEvent<HTMLFormElement>) {
+  function handleAuthTabKeyDown(event: KeyboardEvent<HTMLButtonElement>) {
+    const nextMethod = event.key === "ArrowLeft" || event.key === "ArrowUp" || event.key === "Home"
+      ? "qr"
+      : event.key === "ArrowRight" || event.key === "ArrowDown" || event.key === "End"
+        ? "credential"
+        : null;
+    if (!nextMethod) return;
     event.preventDefault();
-    if (!consentReady || !credential.trim() || importState === "submitting") return;
-    setImportState("submitting");
-    setImportError(null);
-    try {
-      const session = await importSklandCredential(credential, policyConsent());
-      if (!mountedRef.current) return;
-      setCredential("");
-      onAuthenticated(session);
-    } catch (error) {
-      if (!mountedRef.current) return;
-      const detail = toDisplayError(error, "凭证导入失败，请稍后重试。");
-      setImportError(`${detail.message}（${detail.code}${detail.requestId ? ` · ${detail.requestId}` : ""}）`);
-    } finally {
-      if (mountedRef.current) setImportState("idle");
-    }
+    setAuthMethod(nextMethod);
+    window.requestAnimationFrame(() => document.getElementById(`${authTabsId}-${nextMethod}-tab`)?.focus());
   }
 
   const pageStatusText = scanState === "loading"
@@ -318,20 +246,38 @@ export function SklandLoginPanel({
             <AlertDescription>{disabledReason ?? "当前未开放森空岛登录，可继续使用 MAA 导入。"}</AlertDescription>
           </Alert>
         ) : (
-          <Tabs value={authMethod} onValueChange={(value) => setAuthMethod(value as AuthMethod)}>
-            <TabsList className="grid w-full grid-cols-2" aria-label="森空岛登录方式">
-              <TabsTrigger value="qr"><ScanLine />扫码登录</TabsTrigger>
-              <TabsTrigger value="credential"><KeyRound />凭证导入</TabsTrigger>
-            </TabsList>
+          <div>
+            <div className="grid min-h-11 w-full grid-cols-2 rounded-lg bg-muted p-1 text-muted-foreground" role="tablist" aria-label="森空岛登录方式">
+              {(["qr", "credential"] as const).map((method) => (
+                <button
+                  key={method}
+                  id={`${authTabsId}-${method}-tab`}
+                  type="button"
+                  role="tab"
+                  aria-selected={authMethod === method}
+                  aria-controls={`${authTabsId}-${method}-panel`}
+                  tabIndex={authMethod === method ? 0 : -1}
+                  className={cn(
+                    "flex min-h-9 items-center justify-center gap-2 rounded-md px-3 text-sm font-medium transition-colors focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-ring",
+                    authMethod === method && "bg-background text-foreground shadow-sm",
+                  )}
+                  onClick={() => setAuthMethod(method)}
+                  onKeyDown={handleAuthTabKeyDown}
+                >
+                  {method === "qr" ? <><ScanLine />扫码登录</> : <><KeyRound />凭证导入</>}
+                </button>
+              ))}
+            </div>
 
-            <TabsContent value="qr" className="mt-4">
+            {authMethod === "qr" ? (
+            <div id={`${authTabsId}-qr-panel`} role="tabpanel" aria-labelledby={`${authTabsId}-qr-tab`} className="mt-4">
               <div className="grid gap-6 md:grid-cols-[minmax(0,1fr)_15rem] md:items-center" data-skland-login-qr>
                 <div className="order-2 grid gap-4 md:order-1">
                   <p className="flex items-start gap-2 text-sm leading-6 text-muted-foreground">
                     <ShieldCheck className="mt-1 size-4 shrink-0 text-emerald-600" aria-hidden="true" />
                     <span>推荐方式。打开森空岛 App 扫描二维码，确认后同步当前角色的干员、基建与状态中心数据。</span>
                   </p>
-                  <PolicyConsent
+                  <SklandPolicyConsent
                     termsAccepted={termsAccepted}
                     privacyAccepted={privacyAccepted}
                     onTermsChange={setTermsAccepted}
@@ -383,102 +329,18 @@ export function SklandLoginPanel({
                   <p className="text-center text-sm leading-6 text-muted-foreground" role="status" aria-live="polite">{pageStatusText}</p>
                 </div>
               </div>
-            </TabsContent>
+            </div>
+            ) : null}
 
-            <TabsContent value="credential" className="mt-4" data-skland-credential-panel>
-              <form className="grid gap-5" onSubmit={(event) => void submitCredential(event)}>
-                <Alert className="md:hidden">
-                  <ScanLine />
-                  <AlertTitle>手机端优先使用扫码</AlertTitle>
-                  <AlertDescription>凭证导入需要桌面浏览器的开发者工具，手机上建议切回“扫码登录”。</AlertDescription>
-                </Alert>
-
-                <ol className="grid gap-4 text-sm leading-6">
-                  <li className="grid grid-cols-[1.75rem_minmax(0,1fr)] gap-3">
-                    <span className="font-number grid size-7 place-items-center rounded-full bg-foreground text-xs font-semibold text-background">1</span>
-                    <div>
-                      <p className="font-medium text-foreground">登录森空岛网页版</p>
-                      <p className="text-muted-foreground">打开 <a className="font-medium text-foreground underline underline-offset-4" href="https://www.skland.com/index" target="_blank" rel="noreferrer">www.skland.com/index</a> 并完成登录。</p>
-                    </div>
-                  </li>
-                  <li className="grid grid-cols-[1.75rem_minmax(0,1fr)] gap-3">
-                    <span className="font-number grid size-7 place-items-center rounded-full bg-foreground text-xs font-semibold text-background">2</span>
-                    <div className="min-w-0">
-                      <p className="font-medium text-foreground">在 F12 Console 执行命令</p>
-                      <p className="text-muted-foreground">打开开发者工具的 Console，将下方命令粘贴并执行。它只读取森空岛页面已有的 cred 与 token，并复制为一行。</p>
-                      <div className="mt-2 flex min-w-0 items-stretch gap-2">
-                        <code className="min-w-0 flex-1 overflow-hidden text-ellipsis whitespace-nowrap rounded-lg bg-muted px-3 py-2 font-number text-xs text-foreground" title={SKLAND_CREDENTIAL_COPY_COMMAND}>{SKLAND_CREDENTIAL_COPY_COMMAND}</code>
-                        <Button type="button" variant="outline" size="sm" onClick={() => void copyCredentialCommand()} data-skland-copy-command>
-                          {copyState === "copied" ? <Check /> : <Clipboard />}
-                          {copyState === "copied" ? "已复制" : "复制命令"}
-                        </Button>
-                      </div>
-                      {copyState === "error" ? <p className="mt-1 text-xs text-destructive">复制失败，请手动选择命令复制。</p> : null}
-                    </div>
-                  </li>
-                  <li className="grid grid-cols-[1.75rem_minmax(0,1fr)] gap-3">
-                    <span className="font-number grid size-7 place-items-center rounded-full bg-foreground text-xs font-semibold text-background">3</span>
-                    <div>
-                      <label className="font-medium text-foreground" htmlFor={credentialInputId}>粘贴生成的单行凭证</label>
-                      <p className="text-muted-foreground">内容应为 <span className="font-number">cred,token</span>，仅在本次请求和当前组件内存中短暂存在。</p>
-                      <Input
-                        id={credentialInputId}
-                        type="password"
-                        value={credential}
-                        onChange={(event) => setCredential(event.target.value)}
-                        placeholder="粘贴 cred,token"
-                        autoComplete="off"
-                        autoCapitalize="none"
-                        autoCorrect="off"
-                        spellCheck={false}
-                        maxLength={12 * 1024}
-                        className="mt-2 h-11 font-number"
-                        aria-invalid={Boolean(importError)}
-                        data-skland-credential-input
-                      />
-                    </div>
-                  </li>
-                </ol>
-
-                <div className="grid gap-3 border-y py-4">
-                  <div className="flex items-start gap-2">
-                    <TriangleAlert className="mt-0.5 size-4 shrink-0 text-amber-600" aria-hidden="true" />
-                    <div>
-                      <p className="font-medium text-foreground">凭证等同登录权限，请勿发送给任何人</p>
-                      <p className="mt-1 text-xs leading-5 text-muted-foreground">泄露后可能被用于访问账号与角色、干员练度、基建排班及仓库物资数量等数据。</p>
-                    </div>
-                  </div>
-                  <ul className="grid grid-cols-2 gap-x-4 gap-y-1 text-xs text-muted-foreground sm:grid-cols-4" aria-label="凭证可能访问的数据">
-                    <li>账号与角色</li>
-                    <li>干员练度</li>
-                    <li>基建与排班</li>
-                    <li>仓库物资数量</li>
-                  </ul>
-                  <p className="text-xs font-medium leading-5 text-foreground">本站实际不读取、不保存、不展示仓库数据，也不签到、不发布或操作任何社区内容。</p>
-                </div>
-
-                <PolicyConsent
-                  termsAccepted={termsAccepted}
-                  privacyAccepted={privacyAccepted}
-                  onTermsChange={setTermsAccepted}
-                  onPrivacyChange={setPrivacyAccepted}
-                />
-
-                {importError ? <Alert variant="destructive"><AlertDescription>{importError}</AlertDescription></Alert> : null}
-
-                <Button
-                  type="submit"
-                  size={dialogPresentation ? "dialog" : "default"}
-                  className={cn("w-full sm:w-fit", dialogPresentation && "sm:w-[196px]")}
-                  disabled={!consentReady || !credential.trim() || importState === "submitting"}
-                  data-skland-credential-submit
-                >
-                  {importState === "submitting" ? <LoaderCircle className="animate-spin motion-reduce:animate-none" /> : <KeyRound />}
-                  {importState === "submitting" ? "正在验证凭证…" : "导入并登录"}
-                </Button>
-              </form>
-            </TabsContent>
-          </Tabs>
+            {authMethod === "credential" ? (
+            <div id={`${authTabsId}-credential-panel`} role="tabpanel" aria-labelledby={`${authTabsId}-credential-tab`} className="mt-4" data-skland-credential-panel>
+              <SklandCredentialPanel
+                dialogPresentation={dialogPresentation}
+                onAuthenticated={onAuthenticated}
+              />
+            </div>
+            ) : null}
+          </div>
         )}
       </div>
     </Card>
